Enso CompilationStage is a general concept that can be shared accross the projects
