+++ engine/runtime-instrument-common/src/main/java/org/enso/interpreter/instrument/job/SerializeModuleJob.java
@@ -4,8 +4,8 @@ import java.util.logging.Level;
 import org.enso.compiler.SerializationManager;
 import org.enso.interpreter.instrument.execution.RuntimeContext;
 import org.enso.interpreter.runtime.EnsoContext;
-import org.enso.interpreter.runtime.Module;
 import org.enso.pkg.QualifiedName;
+import org.enso.polyglot.CompilationStage;
 import org.enso.polyglot.RuntimeOptions;
 
 /** The job that serializes module. */
@@ -36,7 +36,7 @@ public final class SerializeModuleJob extends BackgroundJob<Void> {
           .findModule(moduleName.toString())
           .ifPresent(
               module -> {
-                if (module.getCompilationStage().isBefore(Module.CompilationStage.AFTER_CODEGEN)) {
+                if (module.getCompilationStage().isBefore(CompilationStage.AFTER_CODEGEN)) {
                   ctx.executionService()
                       .getLogger()
                       .log(
