+++ engine/runtime/src/main/java/org/enso/interpreter/runtime/Module.java
@@ -40,6 +40,7 @@ import org.enso.interpreter.runtime.scope.ModuleScope;
 import org.enso.interpreter.runtime.type.Types;
 import org.enso.pkg.Package;
 import org.enso.pkg.QualifiedName;
+import org.enso.polyglot.CompilationStage;
 import org.enso.polyglot.LanguageInfo;
 import org.enso.polyglot.MethodNames;
 import org.enso.text.buffer.Rope;
@@ -48,44 +49,6 @@ import scala.Function1;
 /** Represents a source module with a known location. */
 @ExportLibrary(InteropLibrary.class)
 public final class Module implements TruffleObject {
-
-  /** Defines a stage of compilation of the module. */
-  public enum CompilationStage {
-    INITIAL(0),
-    AFTER_PARSING(1),
-    AFTER_IMPORT_RESOLUTION(2),
-    AFTER_GLOBAL_TYPES(3),
-    AFTER_STATIC_PASSES(4),
-    AFTER_RUNTIME_STUBS(5),
-    AFTER_CODEGEN(6);
-
-    private final int ordinal;
-
-    CompilationStage(int ordinal) {
-      this.ordinal = ordinal;
-    }
-
-    /**
-     * Checks whether the current compilation stage is at least as advanced as the provided one.
-     *
-     * @param stage the stage to compare to.
-     * @return whether or not {@code this} is at least as advanced as {@code stage}.
-     */
-    public boolean isAtLeast(CompilationStage stage) {
-      return ordinal >= stage.ordinal;
-    }
-
-    /**
-     * Checks that the current compilation stage is before the provided one.
-     *
-     * @param stage the stage to compare to.
-     * @return whether or not {@code this} is before then {@code stage}.
-     */
-    public boolean isBefore(CompilationStage stage) {
-      return ordinal < stage.ordinal;
-    }
-  }
-
   private ModuleScope scope;
   private ModuleSources sources;
   private PatchedModuleValues patchedValues;
