+++ engine/runtime/src/main/scala/org/enso/compiler/phase/ImportResolver.scala
@@ -14,7 +14,7 @@ import org.enso.compiler.core.CompilerError
 import org.enso.compiler.pass.analyse.BindingAnalysis
 import org.enso.editions.LibraryName
 import org.enso.interpreter.runtime.Module
-
+import org.enso.polyglot.CompilationStage
 import scala.collection.mutable
 
 /** Runs imports resolution. Starts from a given module and then recursively
@@ -23,7 +23,7 @@ import scala.collection.mutable
   * Each of the reachable modules will be parsed and will have imported modules
   * injected into its metadata. In effect, running this will bring every module
   * that could ever be necessary for the entry point compilation to at least
-  * the [[Module.CompilationStage.AFTER_IMPORT_RESOLUTION]] stage.
+  * the [[CompilationStage.AFTER_IMPORT_RESOLUTION]] stage.
   *
   * @param compiler the compiler instance for the compiling context.
   */
@@ -51,7 +51,7 @@ class ImportResolver(compiler: Compiler) {
       if (
         current.getCompilationStage
           .isBefore(
-            Module.CompilationStage.AFTER_IMPORT_RESOLUTION
+            CompilationStage.AFTER_IMPORT_RESOLUTION
           ) || !current.hasCrossModuleLinks
       ) {
         val importedModules: List[
@@ -67,7 +67,7 @@ class ImportResolver(compiler: Compiler) {
         current.unsafeSetIr(newIr)
         if (!current.wasLoadedFromCache()) {
           current.unsafeSetCompilationStage(
-            Module.CompilationStage.AFTER_IMPORT_RESOLUTION
+            CompilationStage.AFTER_IMPORT_RESOLUTION
           )
         }
       }
